requirements.txt
customtkinter==5.2.2
Pillow==11.3.0
psutil==7.0.0
WMI==1.5.1
loguru==0.7.2
numpy==2.3.3
pandas==2.3.2
scikit-learn==1.7.2
scipy==1.16.2
python-dateutil==2.9.0.post0
pywin32==311
tzdata==2025.2
threadpoolctl==3.6.0
joblib==1.5.2
psutil
wmi
requests
nvidia-ml-py3